requirements.txt
Flask==3.0.0
huggingface-hub==0.20.0
python-dotenv==1.0.0
requests==2.31.0